_archive/utility-scripts/requirements.txt
# Python dependencies for DesiTV automation scripts
pymongo>=4.0.0
python-dotenv>=1.0.0
